requirements.txt
python-docx==0.8.11
python-telegram-bot==13.10
